Reduce omni radius of Aeon ACU Sensors (#6745)
Make fire beetles more viable against Aeon by reducing the omni range of the sensors upgrade so that the ACU has to stay at the front to use the omni effectively, which allows counterplay since the ACU and army are possible to threaten. The range is enough to match gun advanced range.

Aeon ACU (UAL0001) Sensor System:

    Omni radius: 80 -> 36